+++ spinetoolbox/project_items/data_store/__init__.py
@@ -22,8 +22,8 @@ from .widgets.data_store_properties_widget import DataStorePropertiesWidget
 from .widgets.add_data_store_widget import AddDataStoreWidget
 
 item_rank = 0
-item_category = "Data Stores"
-item_type = "Data Store"
+item_category = item_maker.category()
+item_type = item_maker.item_type()
 item_icon = ":/icons/project_item_icons/database.svg"
 icon_maker = DataStoreIcon
 properties_widget_maker = DataStorePropertiesWidget